refactor: PostLikeService 공통 interface로 빼기
