+++ src/test/java/com/dku/council/domain/like/service/RedisPostLikeServiceImplTest.java
@@ -5,6 +5,7 @@ import com.dku.council.domain.like.model.LikeState;
 import com.dku.council.domain.like.model.entity.PostLike;
 import com.dku.council.domain.like.repository.PostLikeMemoryRepository;
 import com.dku.council.domain.like.repository.PostLikePersistenceRepository;
+import com.dku.council.domain.like.service.impl.RedisPostLikeServiceImpl;
 import com.dku.council.domain.post.repository.PostRepository;
 import com.dku.council.domain.user.repository.UserRepository;
 import com.dku.council.mock.NewsMock;
@@ -28,7 +29,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class PostLikeServiceTest {
+class RedisPostLikeServiceImplTest {
 
     @Mock
     private PostLikeMemoryRepository memoryRepository;
@@ -43,7 +44,7 @@ class PostLikeServiceTest {
     private PostRepository postRepository;
 
     @InjectMocks
-    private PostLikeService service;
+    private RedisPostLikeServiceImpl service;
 
     @Test
     @DisplayName("좋아요 - 캐시에 count가 없는 경우")
